pyproject.toml
[project]
name = "emotonalai"
version = "0.1.0"
description = "AI that behaves as if it has feelings."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.3",
    "gtts>=2.5.3",
    "numpy>=2.1.2",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.10.4",
]
